Aksel-PPO (#18)
* look at car trajectories, compare against nearest heuristic

* final moving weather

* ppo clip vs policy gradient

* Aksel-PPO-Model-Init
Added:
- ppo-vs-heuristic.ipynb
- __init__.py
- ppo_model.py
Modified:
- heuristic_simple.py
- supply_demand_df.py
- vehicles.py
- cost_reward.py
- df_utils.py
- action_mixin.py
- transition_mixin.py
- notebook_imports.py
- generate_obs_state.py

* Aksel-PPO-Model-cont
Modified:
- scratch.ipynb
- ppo-vs-heuristic.ipynb
- __init__.py
- ppo_model.py

* Aksel-PPO-Model-dim
Modified:
- ppo-vs-heuristic.ipynb
- ppo_model.py

* Aksel-PPO-Model-dim
Modified:
- ppo_model.py

* Aksel-PPO-Model-dim
Modified:
- ppo_model.py

* Dtype-Enforcement
Modified:
- gen_customers.ipynb
- scratch.ipynb
- heuristic-train.ipynb
- ppo-vs-heuristic.ipynb
- active_rides.py
- customer.py
- enriched_df_base.py
- requests.py
- space_dicts.py
- supply_demand_df.py
- vehicles.py
- ENV.py
- obs_space_mixin.py
- transition_mixin.py
- ppo_model.py
- notebook_imports.py

* torch-np-bug-fix
Modified:
- ppo-vs-heuristic.ipynb
- enriched_df_base.py
- ENV.py
- action_mixin.py
- obs_space_mixin.py
- ppo_model.py
- notebook_imports.py

* torch-warning-fix
Modified:
- scratch.ipynb
- ppo-vs-heuristic.ipynb
- obs_space_mixin.py
- ppo_model.py

* stable-to-numpy-not-working
Modified:
- scratch.ipynb
- ppo-vs-heuristic.ipynb
- customer.py
- enriched_df_base.py
- ENV.py
- action_mixin.py
- transition_mixin.py
- ppo_model.py

* stable-to-numpy-not-working
Modified:
- scratch.ipynb
- ppo-vs-heuristic.ipynb
- customer.py
- enriched_df_base.py
- requests.py
- vehicles.py
- action_mixin.py
- transition_mixin.py

* numpy-conversion-working
Modified:
- scratch.ipynb
- ppo-vs-heuristic.ipynb
- transition_mixin.py
- euclidean_L2_embed.py

* numpy-conversion-working
Modified:
- scratch.ipynb
- ppo-vs-heuristic.ipynb
- enriched_df_base.py
- obs_space_mixin.py
- transition_mixin.py
- ppo_model.py

* early-stopping
Added:
- train-ppo.ipynb
Modified:
- ppo-vs-heuristic.ipynb
- ppo_model.py

* eval-pipeline-setup
Added:
- ppo_model_final.pt
- evaluate.py
Modified:
- ppo-vs-heuristic.ipynb
- run-heuristic.ipynb
- train-ppo.ipynb
- ppo_model.py
- notebook_imports.py

* eval-pipeline-setup
Modified:
- ppo-vs-heuristic.ipynb

* model-debug
Added:
- test_min_eval.py
Modified:
- train-ppo.ipynb
- graph_mixin.py
- euclidean_L2_embed.py

* model-debug
Added:
- train_utils.py
Modified:
- ppo-vs-heuristic.ipynb
- train-ppo.ipynb
- evaluate.py
- ppo_model.py
- test_min_eval.py

* model-debug
Modified:
- scratch.ipynb
- run-heuristic.ipynb
- train-ppo.ipynb
- config.py
- ppo_model.py
- train_utils.py
- inspect.py
- post_process_enrichment.py

* model-debug
Modified:
- scratch.ipynb
- run-heuristic.ipynb
- post_process_enrichment.py

* unit-test
Modified:
- run-heuristic.ipynb
- obs_space_mixin.py
- test_min_eval.py

* nan-gradients-not-working
Modified:
- ppo_model_final.pt
- train-ppo.ipynb
- ppo_model.py
- test_min_eval.py

* split-up-functions
Added:
- debug-ppo.ipynb
- pricing_model.py
- train_ppo.py
Modified:
- scratch.ipynb
- ppo-vs-heuristic.ipynb
- ppo_model.py
- train_utils.py
- notebook_imports.py

* sub-actor

* sub-actor

* pricing-sub-actor-init
Added:
- torch_np_utils.py
Modified:
- debug-ppo.ipynb
- ppo_model.py
- pricing_model.py
- sub_actor.py
- test_min_eval.py
- train_ppo.py
- notebook_imports.py

* pricing-sub-actor-init
Modified:
- debug-ppo.ipynb
- transition_mixin.py
- evaluate.py
- ppo_model.py
- pricing_model.py
- test_min_eval.py
- torch_np_utils.py
- train_utils.py

* reposition-dispatch-actors
Added:
- submodel_dispatch.py
- submodel_reposition.py
Modified:
- debug-ppo.ipynb
- ppo_model.py
- train_ppo.py

* small-update
Modified:
- debug-ppo.ipynb
- transition_mixin.py

* small-update
Modified:
- debug-ppo.ipynb
- ppo_model.py

* ppo-running
Added:
- metrics.py
- train_debug.py
Modified:
- ppo_model_final.pt
- debug-ppo.ipynb
- heuristic-depug.ipynb
- train-ppo.ipynb
- customer.py
- requests.py
- space_dicts.py
- vehicles.py
- ENV.py
- df_utils.py
- transition_mixin.py
- evaluate.py
- ppo_model.py
- submodel_base.py
- torch_np_utils.py
- train_ppo.py
- train_utils.py

* ppo-running
Added:
- ppo_model_V-0.80_L-0.01_H-256_NV-50_MR-100.pt
Modified:
- ppo-vs-heuristic.ipynb
- train-ppo.ipynb
- train_utils.py

* tqdm-improvements
Modified:
- ppo_model_V-0.80_L-0.01_H-256_NV-50_MR-100.pt
- train-ppo.ipynb
- train_ppo.py
- train_utils.py

* aksel-ppo
Added:
- ppo_model_V-0.08_L-0.01_H-256_NV-50_MR-100.pt
- ppo_model_V-0.80_L-0.01_H-128_NV-50_MR-100.pt
Modified:
- debug-ppo.ipynb
- ppo-vs-heuristic.ipynb
- train-ppo.ipynb
- config.py
- ppo_model.py
- torch_np_utils.py
- train_ppo.py
- train_utils.py

---------

+++ waymo_agent/osmnx/euclidean_L2_embed.py
@@ -1,7 +1,7 @@
 import networkx as nx
 import numpy as np
 import osmnx as ox
-from scipy.spatial.kdtree import cKDTree
+from scipy.spatial import cKDTree  # type: ignore
 
 PRECISION = 6
 import typing as t
@@ -89,7 +89,7 @@ def interpolate_position_on_edge(
         interpolated = src + perc_done * (tgt - src)
         ret_df[val] = interpolated
 
-    return ret_df
+    return ret_df.astype(np.float32)
 
 
 def f_edges_start_end_node(
